Plugin/markitdown-aspose-cells-plugin/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markitdown-aspose-cells-plugin"
dynamic = ["version"]
description = 'A XLSX plugin via Aspose.Cells for the "markitdown" library.'
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Aspose Cells"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "markitdown>=0.1.0a1",
  "aspose-cells-python",
]

[project.urls]
Documentation = "https://github.com/microsoft/markitdown#readme"
Issues = "https://github.com/microsoft/markitdown/issues"
Source = "https://github.com/microsoft/markitdown"

[tool.hatch.version]
path = "src/markitdown_aspose_cells_plugin/__about__.py"

# IMPORTANT: MarkItDown will look for this entry point to find the plugin.
[project.entry-points."markitdown.plugin"]
aspose_cells_plugin = "markitdown_aspose_cells_plugin"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/markitdown_aspose_cells_plugin tests}"

[tool.coverage.run]
source_pkgs = ["markitdown-aspose-cells-plugin", "tests"]
branch = true
parallel = true
omit = [
  "src/markitdown_aspose_cells_plugin/__about__.py",
]

[tool.coverage.paths]
markitdown-aspose-cells-plugin = ["src/markitdown_aspose_cells_plugin", "*/markitdown-aspose-cells-plugin/src/markitdown_aspose_cells_plugin"]
tests = ["tests", "*/markitdown-aspose-cells-plugin/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.sdist]
only-include = ["src/markitdown_aspose_cells_plugin"]
